Activate the event system
